requirements.txt
# python3 -m venv venv
# source venv/bin/activate
# pip install -r requirements.txt

# python -v
# Python 3.12.3 (main, Jan 17 2025, 18:03:48)

# python3.12 -m venv venv
# pip install --upgrade pip setuptools wheel

# =====
# pip freeze > requirements_1.txt
# Этот метод создаст файл со всеми установленными пакетами и их точными версиями.
# ===

# Требования для калибровки камеры и обработки изображений

# Основные библиотеки
numpy>=1.21.0
opencv-python>=4.5.0

# Для визуализации результатов
matplotlib>=3.5.0

# Для работы с файлами и путями
Pillow>=9.0.0

# Для отображения прогресса (опционально)
tqdm>=4.64.0

# Для работы с YAML конфигурацией
PyYAML>=6.0

# === ???? =====
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
kiwisolver==1.4.9
packaging==26.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
six==1.17.0

